task 2 rezolvat

+++ headers/liste.h
@@ -1,2 +1,2 @@
 void addAtBeginning(Node **, teamInfo, FILE *);
-void deleteNode(Node **, int);
+void deleteNode(Node **, float);
